Add files via upload
alert__demo_data.txt was taken from the repo of the assert paper. Project 3_No dummies.ipynb is the code I used to clean the data and organize it such that the model could read it. This model trains slowly and gets 93% accuracy. A GNN model would be preferable. The goal of the project was to take in IP data or suricata alerts and triage in attack chains in real time. Currently this is a MVP. More work is needed.